Split out system_menu view
